User now has access to a TextArea for better input, including the ability to make new lines and expand the box
This commit is functional but ugly. I cannot figure out why there is a default padding to this component even if i explicity tell it to remove it. For some reason the text itself has ALWAYS more than one line worth of height so having nothing written still makes the component scrollable even if it has enough space. It will also not update height imediatly if you delete the input and instead will slowly go to the expected height with each update. I will come back to this at a later ocmmit